Improve sidebar handling on desktop.
Previously, the logic needed to move the sidebar into view when in the desktop was attached to a load event handler, that we listen to on window. However, that prevents this logic to be seen by Svelte during SSR and so the desktop version had no sidebars when the site was exported. Now this logic has been moved to the onMount lifecycle function of _layout.